+++ lib/src/_utils.dart
@@ -142,11 +142,3 @@ extension StreamUtils<T> on Stream<T> {
     yield* next;
   }
 }
-
-extension IterableUtils<T> on Iterable<T?> {
-  Iterable<T> whereNotNull() sync* {
-    for (final item in this) {
-      if (item != null) yield item;
-    }
-  }
-}